% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_questionnaires.R
\name{clean_likert}
\alias{clean_likert}
\title{Convert Likert items from text to number}
\usage{
clean_likert(vector_in = NULL, text_finish = FALSE,
  exclude_middle = FALSE, full_caution = FALSE)
}
\arguments{
\item{vector_in}{vector of responses to Likert item; default NULL.
should be factor or (ideally) character.  Using a factor vector will
return a message but will still work; any other type except characters
will halt with an error.}

\item{text_finish}{logical; default FALSE.  If TRUE, will return abbreviated
text instead of integers.}

\item{exclude_middle}{logical; default FALSE.  Set to TRUE if working with
a 4-point Likert item.  Not yet implemented.}

\item{full_caution}{logical; default FALSE.  If TRUE, will halt and throw an
an error if there's any values that can't be converted.}
}
\value{
integer vector
}
\description{
Returns an integer vector (1--5) if given text responses to Likert items.
}
\details{
Converts text responses to Likert items to integer responses.  Requires a
character or factor vector as an input; maps this onto the 1-5 scale to
create the output vector.  Will work for 5- and 4-point Likert items.
}
