% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uxer-package.R
\docType{package}
\name{uxer-package}
\alias{uxer}
\alias{uxer-package}
\title{uxer: User Experience Research Functions}
\description{
Helper functions related to user experience research.  Will
    contain functions related to various common questionnaire scales, as well as
    creating graphs that are used fairly often.
}
\author{
\strong{Maintainer}: Adam George Dunn \email{agd1983@gmail.com}

}
\keyword{internal}
