% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_questionnaires.R
\name{score_sus}
\alias{score_sus}
\title{Calculate participants' SUS questionnaire scores.}
\usage{
score_sus(myData, user_id = TRUE)
}
\arguments{
\item{myData}{no default; data frame of questionnaire results; see example
for naming convention expected.}

\item{user_id}{default TRUE; does the data frame have a column for user ID?}
}
\description{
Calculate the SUS score per questionnaire participant.  At the
moment, this function does not rescale to 0-100.
}
\concept{questionnaire, scale
score_sus()}
