% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_questionnaires.R
\name{score_sus}
\alias{score_sus}
\title{Calculate participants' SUS questionnaire scores}
\usage{
score_sus(myData, user_id = TRUE, rescale = TRUE, subscales = FALSE)
}
\arguments{
\item{myData}{no default; data frame of questionnaire results.  See example
for naming convention expected.}

\item{user_id}{default TRUE; does the data frame have a column for user ID?}

\item{rescale}{default TRUE; transform main scale and subscales to 0-100
range?}

\item{subscales}{default FALSE; calculate subscales `Usable' and `Learnable'?}
}
\description{
Calculate the SUS score per questionnaire participant.  By default, will
  rescale it to 0-100 range.  Also calculates the `usable' and `learnable'
  subscales if asked to.
}
\references{
there should be a reference here
}
\concept{questionnaire, scale}
