% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_questionnaires.R
\name{score_tam}
\alias{score_tam}
\title{Calculate participants' scores on technology acceptance model}
\usage{
score_tam(myData, user_id = TRUE, usef_start = user_id + 1,
  ease_start = user_id + 7)
}
\arguments{
\item{myData}{a data frame with (optional) participant ID,
then (mandatory) usefulness scale questions followed by
(mandatory) ease-of-use scale questions.}

\item{user_id}{boolean for presence or absence of user ID column as first
column.}

\item{usef_start}{left-most column of usefulness scale, default is 1 
higher than user_id.}

\item{ease_start}{left-most column of ease-of-use scale, default is 7 higher
than user_id.}
}
\description{
Calculate participants' scores on full technology acceptance model.  There
  are two subscales: `perceived usefulness' and `perceived ease of use'.
  Currently assumes participant ID is first column (if present),
  then usefulness scale, then ease scale.
  It's based just off the original paper and no later extensions (yet).
}
\references{
Davis, Fred D. (1989) ``Perceived Usefulness, Perceived Ease
  of Use, and User Acceptance of Information'' \emph{MIS Quarterly} 13 (3)
}
\concept{questionnaire, scale}
