% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_questionnaires.R
\name{score_tam}
\alias{score_tam}
\title{Calculate participants' scores on technology acceptance model}
\usage{
score_tam(myData)
}
\description{
Calculate participants' scores on full technology acceptance model.
}
\references{
Davis, Fred D. (1989) ``Perceived Usefulness, Perceived Ease
  of Use, and User Acceptance of Information'' \emph{MIS Quarterly} 13 (3)
}
\concept{questionnaire, scale}
