% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_study}
\alias{small_study}
\title{Responses to an SUS questionnaire.}
\format{A tibble of 8 rows and 11 variables.
\describe{
  \item{participant}{ID variable for participant in study.}
  \item{Q1 to Q10}{responses to standard SUS questions, range 1--5}
}}
\source{
This is fictional data.
}
\usage{
small_study
}
\description{
8 fictitious participants responded to a standard SUS questionnaire
as part of usability testing a made-up software package.  Nothing 
sophisticated was done to imagine up these data.
}
\keyword{datasets}
